backend/authorization-service/requirements.txt
# WEB FRAMEWORK
aiohttp>=3.8.5
aiohttp-cors>=0.7.0
asyncio>=3.4.3

# DATABASE
asyncpg>=0.27.0

# AUTHENTICATION
PyJWT==2.7.0
cryptography==40.0.1

# EMAIL
aiosmtplib==2.0.1
Jinja2==3.1.2

# UTILITIES
python-dotenv==1.0.0

# PROMETHEUS TRACING
prometheus-client>=0.16.0

# JAEGER TRACING
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
opentelemetry-exporter-jaeger>=1.18.0
opentelemetry-instrumentation-aiohttp-client>=0.40b0
opentelemetry-instrumentation-asyncpg>=0.40b0